requirements.txt
requests>=2.31
pydantic>=2.0
PyYAML>=6.0
beautifulsoup4>=4.12
lxml>=5.0
python-dotenv>=1.0
pytest-timeout>=2.3
